修改protobuf协议，标记消息类型，第一位表示协议版本，从1号开始；第二位表示业务消息或IM消息。

+++ TerminalSDK4Android/src/main/java/ptt/terminalsdk/manager/channel/ClientChannel.java
@@ -64,7 +64,7 @@ public class ClientChannel extends AbsClientChannel {
 	}
 
 	@Override
-	public void sendMessage(GeneratedMessage message, final PushMessageSendResultHandler handler, int messageFun) {
+	public void sendMessage(GeneratedMessage message, final PushMessageSendResultHandler handler, byte messageFun) {
 		if(MyTerminalFactory.getSDK().hasNetwork()){
 			logger.info("发送消息："+message.getClass()+"----->"+message);
 			try {
